requirements.txt
langchain==0.1.0
langchain-mistralai==0.0.3
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.32.0